refactor: auth 관련 파일 및 디렉토리의 이름을 account로 변경
- Auth 기능과 Account 엔티티 사이를 비교해 Auth에 더 가까운 것의 이름은 유지
