+++ src/main/java/com/ss6051/backendspring/account/AccountRepository.java
@@ -1,4 +1,4 @@
-package com.ss6051.backendspring.auth;
+package com.ss6051.backendspring.account;
 
 import com.ss6051.backendspring.global.domain.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
